Got it finally working in a basic setup
Using webpack package with ts-loader to get it working. Not
sure if this is preferable to TSC compiling and gulping and
all other such sourcery but this will do for now. More
investigation later

Added typings too, based on following a tutorial on the
internetz. Not sure if we'll need it, so it may come out
later