Fixed panning on some sounds... sort of
Because we now move the POV, we can't just set an absolute
panning from the POV, so we have to pan the actual sounds
in relation to the POV for some reason. So do that with
the bump sounds. It's probably too much to do it for the
running sounds for little gain